requirements.txt
# PPT编辑器依赖包

# OCR识别（CPU版本，如需GPU请安装paddlepaddle-gpu）
paddleocr>=2.7.0
paddlepaddle>=2.6.0

# PPT生成
python-pptx>=0.6.21

# 图像处理
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# PDF支持
PyMuPDF>=1.23.0
